webkit/data/layout_tests/platform/chromium-win/LayoutTests/fast/events/drag-in-frames-expected.txt

Event log

ondragstart src
ondrag src
ondragenter left target
ondrag src
ondragover left target
ondrag src
ondragenter right target
ondragleave left target
ondrag src
ondragover right target
ondragover right target
ondragend src
ondrop right target
These results are just "sensible" at the moment, meaning that all the events fire on the right elements, nothing more. The order in particular can rightfully change.
